requirements.txt
aiohttp
aiofiles
types-aiofiles
python-dotenv